SECTION 17/Simple_UI_new/requirement.txt
streamlit>=1.35.0
gradio>=4.0.0
pandas>=2.0.0
numpy>=1.26.0,<2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
Pillow>=10.0.0
